added script to the tool, tool FUNCTIONS
the tool now works in my ideal version, if more components of different types need to be added, then variations of lines 92-117 must be added. apparently, different types need to be specified in order for the script to access different types of components from folders in the root folder fields.